\name{Missing}
\alias{Missing}
\docType{data}
\title{
	Missing gene expression data
}
\description{
	M by N matrix corresponding to elements of Y. 0 denotes no missing, while 1 denotes missing
}
\usage{data(Missing)}
\format{
  The format is:
 num [1:30, 1:200] 0 0 0 0 0 0 0 0 0 0 ...
}
\references{
1. Cai, X., Bazerque, J.A., and Giannakis, G.B. (2013). Inference of Gene Regulatory Networks with Sparse Structural Equation Models Exploiting Genetic Perturbations. PLoS Comput Biol 9, e1003068. \cr
2. Huang A., Bazerque J. A., Giannakis G. B., DUroso G., Myers C. L., Cai X., Elastic Net algorithm for inferring gene regulatory networks based on structural equation models, to be submitted. \cr
}
\examples{
data(Missing)
}
\keyword{datasets}
