\name{sparseSEM-internal}
\title{Internal sparseSEM function}
\alias{enSEM_STS}

\description{Internal sparseSEM function for Stability Selection}
\usage{
	enSEM_STS(i,Y,X,Missing ,B,STS_para,kFold,verbose);
	
}
\author{Anhui Huang}

\details{These are not intended for use by users. 
	\code{enSEM_STS} fits sparseSEM model by calling internal C function. 

}
\keyword{internal}
