\name{Y}
\alias{Y}
\docType{data}
\title{
	Gene expression matrix
}
\description{
	Y is the M by N matrix describes the expression profile of M genes in N samples
}
\usage{data(Y)}
\format{
  The format is:
 num [1:30, 1:200] 3.02 1.12 -2.24 3.58 2.18 ...
}
\details{
	Gene expression data
}
\references{
1. Cai, X., Bazerque, J.A., and Giannakis, G.B. (2013). Inference of Gene Regulatory Networks with Sparse Structural Equation Models Exploiting Genetic Perturbations. PLoS Comput Biol 9, e1003068. \cr
2. Huang A., Bazerque J. A., Giannakis G. B., DUroso G., Myers C. L., Cai X., Elastic Net algorithm for inferring gene regulatory networks based on structural equation models, to be submitted. \cr
}
\examples{
data(Y)
}
\keyword{datasets}
