\name{sparseSEM-package}
\alias{sparseSEM-package}
\alias{sparseSEM}
\docType{package}
\title{
	sparseSEM: Sparse-Aware Maximum Likelihood for Structural Equation Models in Inferring Gene Regulatory Networks
}
\description{
	state-of-the-art sparse aware maximum likelihood function for structural equation models.
	Two penalty functions including the Lasso and Elastic net available.
	Version 3.8-1: add output for cross validation k-fold mean and ste
				
				
}
\details{
\tabular{ll}{
Package: \tab sparseSEM\cr
Type: \tab Package\cr
Version: \tab 3.2\cr
Date: \tab 2023-04-17\cr
License: \tab GPL\cr
}

}
\author{
Anhui Huang

Maintainer: Anhui Huang <anhuihuang@gmail.com>

}
\references{
1. Cai, X., Bazerque, J.A., and Giannakis, G.B. (2013). Inference of Gene Regulatory Networks with Sparse Structural Equation Models Exploiting Genetic Perturbations. PLoS Comput Biol 9, e1003068. \cr
2. Huang, A. (2014). "Sparse model learning for inferring genotype and phenotype associations." Ph.D Dissertation. University of Miami(1186).\cr \cr	 \cr
}
\keyword{structural equation models}

\examples{
library(sparseSEM)
}
